Success?

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -8,8 +8,8 @@
  */
 int get_endianness(void)
 {
-	int i = 1;
+	unsigned int i = 1;
 	char *c = (char *)&i;
 
-	return (*c);
+	return ((int)*c);
 }
